requirements.txt
streamlit
pandas
google-generativeai
python-dotenv
gspread
oauth2client